+++ docs/source/whats_new.rst
@@ -28,6 +28,7 @@ Enhancements
 - Updated Unit Tests (:gh:`14` by `Stella Douka`_)
 - Option to disable logging (:gh:`14` by `Stella Douka`_)
 - Add CI (:gh:`2` by `Sylvain Chevallier`_)
+- Modify LinearGrowingModule to operate on the last dimension of an input tensor with arbitrary shape (:gh:`54` by `Stéphane Rivaud`_)
 
 Bugs
 ~~~~